+++ .gitignore
@@ -7,3 +7,4 @@ server-dest
 dest/config/ 
 dest/routes/ 
 dest/app.js
+dest/controllers/
