+++ src/routes/authRoutes.js
@@ -1,20 +1,14 @@
 'use strict'; 
 const express = require('express');
-const { MongoClient } = require('mongodb'); 
-const debug = require('debug')('app:authRoutes');
+const authController = require('../controllers/authController'); 
 const passport = require('passport'); 
-const { body, validationResult } = require('express-validator/check'); 
-
+const { body } = require('express-validator/check'); 
 const authRouter = express.Router(); 
 
 function router() {
+  const { getRegister, postRegister, getLogin, getProfile, getLogout } = authController(); 
   authRouter.route('/register')
-    .get((req, res) => {
-      res.render('register', {
-        title: 'Register',
-        errors: []
-      });
-    }) 
+    .get(getRegister) 
     .post([
       body('username', 'Empty Username Field').not().isEmpty().trim().escape(), 
       body('password', 'Password must be at least 5 characters').isLength({ min: 5})
@@ -25,71 +19,18 @@ function router() {
                 return value;
               }
             })
-      ], (req, res) => {
-        const errors = validationResult(req);
-        if (!errors.isEmpty()) {
-          const update = errors.array(); 
-          console.log(update); 
-          return res.render('register', {
-            title: 'Register',
-            errors: update 
-          }); 
-        }
-        const { username, password } = req.body; 
-        const url = 'mongodb://localhost:27017'; 
-        const dbName = 'parodyApp'; 
-
-        (async function addUser() {
-          let client;
-          try {
-            client = await MongoClient.connect(url); 
-            const db = client.db(dbName); 
-            const col = db.collection('users'); 
-            const user = { username, password };
-            const exists = await col.findOne({username: user.username}); 
-            console.log(exists);  
-            if (exists === null) {
-              const results = await col.insertOne(user);
-              req.login(results.ops[0], () => {
-                res.redirect('/auth/profile');
-              });} else {
-                return res.render('register', {
-                  title: 'Register',
-                  errors: [{msg: 'Username already taken!'}]}); 
-              } 
-          } catch (err) {
-            debug(err); 
-          }
-          }()); 
-      }); 
+      ], postRegister); 
   authRouter.route('/login')
-   .get((req, res) => {
-     res.render('login', {
-       title: 'Login', 
-       message: req.flash('error')
-      }); 
-    }) 
+   .get(getLogin) 
    .post(passport.authenticate('local', {
      successRedirect: '/auth/profile',
      failureRedirect: '/auth/login',
      failureFlash: true
     }));  
   authRouter.route('/profile')
-    .get((req, res) => {
-      if (req.user) {
-        res.render('profile', {
-          title: 'Profile', 
-          username: req.user.username
-        }); 
-      } else {
-        res.redirect('/'); 
-      }
-    });  
+    .get(getProfile);  
   authRouter.route('/logout')
-    .get((req, res) => {
-      req.logout();
-      res.redirect('/'); 
-    }); 
+    .get(getLogout); 
   return authRouter; 
   }
